refactor(governance): 清理大文件基线硬债
