+++ docs/architecture/large-file-baseline.json
@@ -1,5 +1,5 @@
 {
-  "generatedAt": "2026-04-22T17:54:39.257Z",
+  "generatedAt": "2026-06-03T07:48:13.592Z",
   "scope": "fail",
   "policyVersion": "2026-04-22.policy-v2",
   "entries": []
