+++ openspec/changes/archive/2026-06-03-eliminate-large-file-baseline-debt/.openspec.yaml
@@ -1,0 +1,2 @@
+schema: spec-driven
+created: 2026-06-03
